[v2.x] skip specs if plugin is missing

+++ lib/chewy/query.rb
@@ -907,18 +907,16 @@ module Chewy
     #   UsersIndex::User.filter{ age <= 42 }.delete_all
     #
     def delete_all
+      if Runtime.version > '2.0'
+        plugins = Chewy.client.nodes.info(plugins: true)["nodes"].values.map { |item| item["plugins"] }.flatten
+        raise PluginMissing, "install delete-by-query plugin" unless plugins.find { |item| item["name"] == 'delete-by-query' }
+      end
       request = chain { criteria.update_options simple: true }.send(:_request)
       ActiveSupport::Notifications.instrument 'delete_query.chewy',
         request: request, indexes: _indexes, types: _types,
         index: _indexes.one? ? _indexes.first : _indexes,
         type: _types.one? ? _types.first : _types do
           Chewy.client.delete_by_query(request)
-      end
-    rescue Elasticsearch::Transport::Transport::Errors::BadRequest => e
-      if e.message =~ /No handler found for uri/
-        raise PluginMissing, "install delete-by-query plugin"
-      else
-        raise e
       end
     end
